Add test for content provider and refactor test logic into UtilsForTest object, and add property to samplePropertyList
