+++ app/build.gradle
@@ -82,7 +82,7 @@ dependencies {
 
     //Glide
     implementation 'com.github.bumptech.glide:glide:4.13.2'
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.12.0'
+    kapt 'com.github.bumptech.glide:compiler:4.12.0'
 
     //Retrofit
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
