+++ app/src/androidTest/java/com/openclassrooms/realestatemanager/ContentProviderTest.kt
@@ -3,6 +3,7 @@ package com.openclassrooms.realestatemanager
 import android.content.ContentResolver
 import android.content.ContentUris.withAppendedId
 import android.content.ContentValues
+import android.database.Cursor
 import android.net.Uri
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -13,6 +14,8 @@ import com.openclassrooms.realestatemanager.model.Property
 import com.openclassrooms.realestatemanager.model.PropertyStatus
 import com.openclassrooms.realestatemanager.provider.RealEstateContentProvider
 import com.openclassrooms.realestatemanager.utils.DummyRealEstateProvider
+import com.openclassrooms.realestatemanager.utils.UtilsForTest.generateProperty
+import com.openclassrooms.realestatemanager.utils.UtilsForTest.getPropertyFromCursor
 import junit.framework.Assert.assertEquals
 import junit.framework.Assert.assertNotNull
 import org.junit.Assert
@@ -35,111 +38,57 @@ class ContentProviderTest {
     }
 
     @Test
+    fun insertPropertyWithSuccess() {
+        val values = generateProperty()
+        val resultUri = contentResolver.insert(uri, values)
+        assertNotNull(resultUri)
+
+        val cursor = contentResolver.query(
+            withAppendedId(
+                RealEstateContentProvider.URI_ITEM,
+                propertyID.toLong()
+            ), null, null, null, null
+        )
+
+        assertNotNull(cursor)
+        assertEquals(1, cursor?.count)
+
+        if (cursor != null && cursor.moveToFirst()) {
+            do {
+                val insertedProperty = getPropertyFromCursor(cursor)
+                val matchingProperty = propertyID.toLong() == insertedProperty.id
+                assertNotNull(matchingProperty)
+            } while (cursor.moveToNext())
+        }
+
+        cursor?.close()
+    }
+
+    @Test
     fun getFirstPropertyOfTableIfExists() {
-        val cursor = contentResolver.query(withAppendedId(uri, propertyID.toLong()), null, null, null, null)
+        val cursor =
+            contentResolver.query(withAppendedId(uri, propertyID.toLong()), null, null, null, null)
         if (cursor!!.count == 1) {
             Assert.assertTrue(cursor.moveToFirst())
-            Assert.assertEquals(propertyID, cursor.getInt(cursor.getColumnIndexOrThrow("property_id")))
+            Assert.assertEquals(
+                propertyID,
+                cursor.getInt(cursor.getColumnIndexOrThrow("property_id"))
+            )
         }
         cursor.close()
     }
 
     @Test
     fun whenFetchingNonExistingPropertyThenCursorIsEmpty() {
-        val cursor = contentResolver.query(withAppendedId(uri, nonExistingPropertyId.toLong()), null, null, null, null)
+        val cursor = contentResolver.query(
+            withAppendedId(uri, nonExistingPropertyId.toLong()),
+            null,
+            null,
+            null,
+            null
+        )
         Assert.assertNotNull(cursor)
         Assert.assertTrue(cursor!!.count == 0)
         cursor.close()
-    }
-
-
-    @Test
-    fun insertValidDataInsertsProperties() {
-        val properties = DummyRealEstateProvider.samplePropertyList
-        val uri1 = Uri.parse("content://com.openclassrooms.realestatemanager.provider/Property/1")
-
-        for (property in properties) {
-            val photosJson = Gson().toJson(property.photos)
-            val values = ContentValues().apply {
-                put("property_id", property.id)
-                put("agent", property.agent)
-                put("type", property.type)
-                put("price", property.price)
-                put("description", property.description)
-                put("address", property.address)
-                put("city", property.city)
-                put("postal_code", property.postalCode)
-                put("country", property.country)
-                put("status", property.status.name)
-                put("number_of_rooms", property.numberOfRooms)
-                put("number_of_bedrooms", property.numberOfBedrooms)
-                put("number_of_bathrooms", property.numberOfBathrooms)
-                put("surface", property.surface)
-                put("point_of_interest", property.pointOfInterest)
-                put("entry_date", property.entryDate.time)
-                put("sale_date", property.saleDate?.time)
-                put("latitude", property.latitude)
-                put("longitude", property.longitude)
-                put("photos", photosJson)
-            }
-            val resultUri = contentResolver.insert(uri1, values)
-            assertNotNull(resultUri)
-        }
-
-        val cursor = contentResolver.query(uri1, null, null, null, null)
-        assertNotNull(cursor)
-        assertEquals(properties.size-1, cursor?.count) //todo can't get the second property
-
-        if (cursor != null) {
-            while (cursor.moveToNext()) {
-                val id = cursor.getLong(cursor.getColumnIndexOrThrow("property_id"))
-                val agent = cursor.getString(cursor.getColumnIndexOrThrow("agent"))
-                val type = cursor.getString(cursor.getColumnIndexOrThrow("type"))
-                val price = cursor.getDouble(cursor.getColumnIndexOrThrow("price"))
-                val description = cursor.getString(cursor.getColumnIndexOrThrow("description"))
-                val address = cursor.getString(cursor.getColumnIndexOrThrow("address"))
-                val city = cursor.getString(cursor.getColumnIndexOrThrow("city"))
-                val postalCode = cursor.getString(cursor.getColumnIndexOrThrow("postal_code"))
-                val country = cursor.getString(cursor.getColumnIndexOrThrow("country"))
-                val status = cursor.getString(cursor.getColumnIndexOrThrow("status"))
-                val numberOfRooms = cursor.getInt(cursor.getColumnIndexOrThrow("number_of_rooms"))
-                val numberOfBedrooms = cursor.getInt(cursor.getColumnIndexOrThrow("number_of_bedrooms"))
-                val numberOfBathrooms = cursor.getInt(cursor.getColumnIndexOrThrow("number_of_bathrooms"))
-                val surface = cursor.getDouble(cursor.getColumnIndexOrThrow("surface"))
-                val pointOfInterest = cursor.getString(cursor.getColumnIndexOrThrow("point_of_interest"))
-                val entryDate = Date(cursor.getLong(cursor.getColumnIndexOrThrow("entry_date")))
-                val saleDate = Date(cursor.getLong(cursor.getColumnIndexOrThrow("sale_date")))
-                val latitude = cursor.getDouble(cursor.getColumnIndexOrThrow("latitude"))
-                val longitude = cursor.getDouble(cursor.getColumnIndexOrThrow("longitude"))
-                val photosJson = cursor.getString(cursor.getColumnIndexOrThrow("photos"))
-                val photos = Gson().fromJson<List<Photo>>(photosJson, object : TypeToken<List<Photo>>() {}.type)
-
-                val insertedProperty = Property(
-                    id = id,
-                    agent = agent,
-                    type = type,
-                    price = price,
-                    description = description,
-                    address = address,
-                    city = city,
-                    postalCode = postalCode,
-                    country = country,
-                    status = PropertyStatus.valueOf(status),
-                    numberOfRooms = numberOfRooms,
-                    numberOfBedrooms = numberOfBedrooms,
-                    numberOfBathrooms = numberOfBathrooms,
-                    surface = surface,
-                    pointOfInterest = pointOfInterest,
-                    entryDate = entryDate,
-                    saleDate = saleDate,
-                    latitude = latitude,
-                    longitude = longitude,
-                    photos = photos
-                )
-                val matchingProperty = properties.find { it.id == insertedProperty.id }
-                assertNotNull(matchingProperty)
-            }
-            cursor.close()
-        }
     }
 }
